Removing clutter from App.js and creating a seperate component for the pokemon list
